ændret budget til int

+++ src/main/java/org/example/eksamenkea/model/Project.java
@@ -4,12 +4,12 @@ package org.example.eksamenkea.model;
 public class Project {
     private int projectId;
     private String projectName;
-    private double budget;
+    private int budget;
     private String projectDescription;
     private int employeeId;
     private int materialCost;
 
-    public Project(int projectId, String projectName, double budget, String projectDescription, int employeeId, int materialCost) {
+    public Project(int projectId, String projectName, int budget, String projectDescription, int employeeId, int materialCost) {
         this.projectId = projectId;
         this.projectName = projectName;
         this.budget = budget;
@@ -18,7 +18,7 @@ public class Project {
         this.materialCost = materialCost;
     }
 
-    public Project(String projectName, double budget, String projectDescription, int employeeId, int materialCost) {
+    public Project(String projectName, int budget, String projectDescription, int employeeId, int materialCost) {
         this.projectName = projectName;
         this.budget = budget;
         this.projectDescription = projectDescription;
@@ -48,7 +48,7 @@ public class Project {
         return budget;
     }
 
-    public void setBudget(double budget) {
+    public void setBudget(int budget) {
         this.budget = budget;
     }
